requirements.txt
pyvmomi
fastapi
uvicorn[standard]
sqlalchemy
sqlalchemy-utils
sqlalchemy-dlock
apscheduler
alembic
dependency-injector
pytz
mysql-connector-python
pymysql